refactor: reduce code duplication in REST API layer (Phase 1)
- Merge CommonResourceOperations v2/v3 into single shared implementation
  with generic transformer function
- Extract shared compareByName() method to ApiDtoUtils
- Update V2ApiUtil and V3ApiUtil to use shared comparison logic

Part of #7305
Closes #7307, #7308
