+++ app/src/main/java/io/apicurio/registry/rest/v2/impl/IdsResourceImpl.java
@@ -12,7 +12,7 @@ import io.apicurio.registry.metrics.health.readiness.ResponseTimeoutReadinessChe
 import io.apicurio.registry.rest.HeadersHack;
 import io.apicurio.registry.rest.v2.IdsResource;
 import io.apicurio.registry.rest.v2.beans.ArtifactReference;
-import io.apicurio.registry.rest.v2.impl.shared.CommonResourceOperations;
+import io.apicurio.registry.rest.impl.shared.CommonResourceOperations;
 import io.apicurio.registry.storage.RegistryStorage;
 import io.apicurio.registry.storage.dto.ArtifactVersionMetaDataDto;
 import io.apicurio.registry.storage.dto.ContentWrapperDto;
@@ -124,7 +124,7 @@ public class IdsResourceImpl implements IdsResource {
      */
     @Override
     public List<ArtifactReference> referencesByContentHash(String contentHash) {
-        return common.getReferencesByContentHash(contentHash);
+        return common.getReferencesByContentHash(contentHash, V2ApiUtil::referenceDtoToReference);
     }
 
     /**
