+++ app/src/main/java/io/apicurio/registry/rest/v3/impl/V3ApiUtil.java
@@ -110,16 +110,8 @@ public final class V3ApiUtil {
 
     public static int compare(SortOrder sortOrder, ArtifactMetaDataDto metaDataDto1,
             ArtifactMetaDataDto metaDataDto2) {
-        String name1 = metaDataDto1.getName();
-        if (name1 == null) {
-            name1 = metaDataDto1.getArtifactId();
-        }
-        String name2 = metaDataDto2.getName();
-        if (name2 == null) {
-            name2 = metaDataDto2.getArtifactId();
-        }
-        return sortOrder == SortOrder.desc ? name2.compareToIgnoreCase(name1)
-            : name1.compareToIgnoreCase(name2);
+        return io.apicurio.registry.rest.ApiDtoUtils.compareByName(
+                sortOrder == SortOrder.desc, metaDataDto1, metaDataDto2);
     }
 
     public static ArtifactSearchResults dtoToSearchResults(ArtifactSearchResultsDto dto) {
